import Darwin normally.
Importing Darwin may be needed by many users of CommandLine, but not all.
If one needs Darwin or GlibC, it is easy enough to add the line to one's own code.

The justification given by the original committer (for adding the @exported attribute) is that it saves one line. That is not a good reason at all.